requirements.txt
alembic==1.5.4
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
cryptography==3.2.1
Cython==0.29.21
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.3
django-allauth==0.43.0
django-autoslug-iplweb==1.9.4
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-sass-processor==0.8.2
django-user-agents==0.4.0
djangorestframework==3.12.2
docupy==0.3.2
docutils==0.16
filelock==3.0.12
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.18.1
idna==2.10
isort==5.5.3
itsdangerous==1.1.0
Jinja2==2.11.3
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.1
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
lazy-object-proxy==1.4.3
libsass==0.20.1
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
mpmath==1.1.0
numpy==1.19.4
nupy==0.1.1
oauthlib==3.1.0
pefile==2019.4.18
pep8==1.7.1
Pillow==6.2.2
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycodestyle==2.6.0
pycparser==2.20
pygame==2.0.1
PyGetWindow==0.0.8
Pygments==2.7.4
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.1
PyMsgBox==1.0.8
PyPDF2==1.26.0
pyperclip==1.8.0
pypiwin32==223
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3
PyQtWebEngine==5.15.2
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
python-editor==1.0.4
python3-openid==3.2.0
pyttsx3==2.90
pytube @ git+https://github.com/pytube/pytube@92856dbf751f09cb46c6282fa9fb2af080dd2418
pytube3==9.6.4
PyTweening==1.0.3
pytz==2020.1
pywin32==300
pywin32-ctypes==0.2.0
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.23
sqlparse==0.3.1
stripe==2.55.0
sympy==1.7.1
toml==0.10.1
tqdm==4.56.0
typing-extensions==3.7.4.3
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
virtualenv==20.0.31
webdrivermanager==0.9.0
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
youtube-dl==2020.11.29
youtube-transcript-api==0.3.1
